requirements.txt
streamlit==1.21.0
pillow==9.5.0
PyMuPDF==1.20.0
numpy==1.23.5
python-dotenv==1.0.0
google-generativeai==0.1.0  # Ensure this is the correct version as per your usage
paddlepaddle==2.5.0  # Verify this version is correct for PaddleOCR
paddleocr==2.5.0  # Ensure this matches with the PaddlePaddle version
